move list of errors from utils to consts
